Improve logging - messaging and Server/Breakpoint classes
Added very detailed logging of messages from/to rdebug-ide and
ruby_debugger.rb and detailed logging of the :Rdebugger command and
toggling breakpoints. They are things which often don't work, so such
detailed logging will help to find/fix user's issues related to the
vim-ruby-debugger plugin.

To enable detailed logging, you need to enable debug_mode. For this,
add

  let g:ruby_debugger_debug_mode = 1

to your .vimrc. It will log to the ruby_debugger_log file (as before).